gpu_server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
mediapipe==0.10.8

# LLM Chatbot dependencies
transformers==4.36.0
peft==0.7.1
accelerate==0.25.0
bitsandbytes==0.41.3
